[Ops][Misc] Refactor 310P W8A8 quantization methods to use a common base class
