+++ vllm_ascend/_310p/quantization/methods/w8a8_dynamic.py
@@ -27,10 +27,11 @@ from vllm_ascend._310p.fused_moe.experts_selector import select_experts
 from vllm_ascend.ascend_forward_context import _EXTRA_CTX
 from vllm_ascend.ops.fused_moe.experts_selector import zero_experts_compute
 from vllm_ascend.ops.fused_moe.moe_runtime_args import build_fused_experts_input
-from vllm_ascend.quantization.methods.base import AscendLinearScheme, AscendMoEScheme, QuantType
+from vllm_ascend.quantization.methods.base import AscendMoEScheme, QuantType
 from vllm_ascend.utils import maybe_trans_nz
 
 from .registry import register_scheme
+from .w8a8_base import AscendW8A8Linear310pScheme
 
 
 @register_scheme("W8A8_DYNAMIC", "moe")
@@ -160,20 +161,12 @@ class AscendW8A8DynamicFusedMoEMethod310(AscendMoEScheme):
 
 
 @register_scheme("W8A8_DYNAMIC", "linear")
-class AscendW8A8DynamicLinearMethod310(AscendLinearScheme):
+class AscendW8A8DynamicLinearMethod310(AscendW8A8Linear310pScheme):
     """310P-only W8A8 dynamic linear scheme.
 
     Notes:
       - This scheme is discovered via 310P local registry.
     """
-
-    def get_weight(
-        self,
-        input_size: int,
-        output_size: int,
-        params_dtype: torch.dtype = torch.float16,
-    ) -> dict[str, Any]:
-        return {"weight": torch.empty(output_size, input_size, dtype=torch.int8)}
 
     def get_perchannel_param(
         self,
