+++ vllm_ascend/_310p/quantization/methods/w8a8sc.py
@@ -23,13 +23,13 @@ import torch_npu
 from vllm.distributed import get_tensor_model_parallel_rank
 
 from vllm_ascend.ops.linear import AscendRowParallelLinear
-from vllm_ascend.quantization.methods.base import AscendLinearScheme
 
 from .registry import register_scheme
+from .w8a8_base import AscendW8A8Linear310pScheme
 
 
 @register_scheme("W8A8SC", "linear")
-class AscendW8A8SCLinearMethod310(AscendLinearScheme):
+class AscendW8A8SCLinearMethod310(AscendW8A8Linear310pScheme):
     """310P-only W8A8SC static linear scheme.
 
     Notes:
@@ -67,18 +67,6 @@ class AscendW8A8SCLinearMethod310(AscendLinearScheme):
             "weight": torch.empty(input_size * output_size, dtype=torch.int8),
             "index": torch.empty(index_len, dtype=torch.int8),
             "info": torch.empty(5, dtype=torch.int64),
-        }
-
-    def get_pertensor_param(self, params_dtype: torch.dtype) -> dict[str, Any]:
-        return {
-            "input_scale": torch.empty(1, dtype=params_dtype),
-            "input_offset": torch.empty(1, dtype=torch.int8),
-        }
-
-    def get_perchannel_param(self, output_size: int, params_dtype: torch.dtype) -> dict[str, Any]:
-        return {
-            "quant_bias": torch.empty(output_size, dtype=torch.int32),
-            "deq_scale": torch.empty(output_size, dtype=torch.int64),
         }
 
     def apply(
